Add markdown support
Add markdown support to improve description rendering.
We often want to add styles in descriptions but currently only text was accepted.
For some case, it was really not enough for everyday use cases.

Also, we add a short description in list views so users can quickly understand what the element (route, resource) is about.
It means changing a little the styles to have something more like a table.

This commit does the following:

* Add `vue-markdown`
* Inject it into `project`, `routes` and `resources` pages where needed
* Add styles for markdown for normal and short description
* Add `Resource.description` when listing all project resources