Add security scheme support
In some application, we need security scheme support.
And some of those security schemes could need custom parameters (eg: with Amazon API Gateway).
So we need to allow those specific attributes.

Once the security scheme is set, we can add it to any existing route.

This commit does the following:

* Add Security Scheme model
* Add it into database
* Add a new menu entry to manage it by project
* A user can create, update and delete a security scheme
* A user can link a security scheme to a route
* Update swagger generation to take security schemes into account