Add API Gateway integration
It is possible to configure Amazon API Gateway with a swagger file.
In some projects, we needed to use this kind of file.
Current swagger file is not enough to do so.

To avoid issues and noises, we do not include those configurations in classical Swagger file.

Also, we needed to be able to configure some parameters (eg: `title`) so the same file would be used in several environments.
We take this opportunity to add specific configuration for this specific file.

This commit does the following:

* Add `api_gateway_integration` in `project`
* Add fields in project form to configure it
* Add a new MIME Type `api_gateway_integration`
* Update `to_swagger` generators to include `api_gateway_integration` or not
* Add a test ensuring the new feature