Add deprecation in routes
Sometimes, the API changes and some routes becomes deprecated.
Currently, we do not have a way to keep the route and indicates it is deprecated.
The only way we have is to update the route description.
However, this information does not appear in generated swagger file.

To explicit this information, we add the attribute `deprecated`.
We show it in route list and route details.

We force by design the user to give an information about the deprecation.

This commit does the following:

* Add `route.deprecated` as string (so the user must give information about it)
* Add the field in route forms
* Show the information in route list
* Show the information in route detail
* Add the information in generated swagger